use billing hook
